# src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabros Bot API",
    "version": "0.1.0",
    "description": "Protected alerting and market-analysis APIs. Machine clients use the x-api-key header; the api-key query parameter exists only for legacy compatibility and may leak through logs. The browser admin console can use Firebase ID tokens on documented admin operations when ENABLE_FIREBASE_ADMIN_AUTH=true."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment"
    }
  ],
...
  "paths": {
...
    "/api/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "List stored alerts",
        "operationId": "listAlerts",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertListLimit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Source"
          },
          {
            "$ref": "#/components/parameters/Enriched"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AlertList"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          },
          {
            "FirebaseBearerAuth": []
          }
        ],
        "x-admin-role": "admin.viewer"
      }
    },
    "/api/alerts/summary": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Summarize stored alert analytics",
        "operationId": "summarizeAlerts",
        "parameters": [
          {
            "$ref": "#/components/parameters/From"
          },
          {
            "$ref": "#/components/parameters/To"
          },
          {
            "$ref": "#/components/parameters/SummaryLimit"
          },
          {
            "$ref": "#/components/parameters/Source"
          },
          {
            "$ref": "#/components/parameters/Enriched"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AlertSummaryResult"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          },
          {
            "FirebaseBearerAuth": []
          }
        ],
        "x-admin-role": "admin.viewer"
      }
    },
    "/api/alerts/export": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Export a bounded alert window",
        "operationId": "exportAlerts",
        "parameters": [
          {
            "$ref": "#/components/parameters/FromRequired"
          },
          {
            "$ref": "#/components/parameters/ToRequired"
          },
          {
            "$ref": "#/components/parameters/ExportFormat"
          },
          {
            "$ref": "#/components/parameters/ExportLimit"
          },
          {
            "$ref": "#/components/parameters/Source"
          },
          {
            "$ref": "#/components/parameters/Enriched"
          },
          {
            "$ref": "#/components/parameters/IncludeText"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON Lines or CSV export",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "type": "string"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          },
          {
            "FirebaseBearerAuth": []
          }
        ],
        "x-admin-role": "admin.viewer"
      }
    },
    "/api/alerts/{alertId}": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get a stored alert",
        "operationId": "getAlertById",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonResult"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          },
          {
            "FirebaseBearerAuth": []
          }
        ],
        "x-admin-role": "admin.viewer"
      }
    },
    "/api/alerts/{alertId}/replay": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "summary": "Replay a stored alert",
        "operationId": "replayAlert",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertId"
          },
          {
            "$ref": "#/components/parameters/DryRun"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          },
          {
            "$ref": "#/components/parameters/XIdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Replay"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/DeliveryResult"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          },
          {
            "FirebaseBearerAuth": []
          }
        ],
        "x-admin-role": "admin.operator"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api-key",
        "description": "Legacy compatibility only. Prefer x-api-key because query strings may be logged."
      },
      "FirebaseBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token for the browser admin console. Requires an admin.viewer or admin.operator custom claim; webhook operations remain API-key-only."
      }
    },
    "parameters": {
      "AlertId": {
        "name": "alertId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "alert-123"
      },
...
      "IdempotencyKeyHeader": {
        "name": "idempotency-key",
        "in": "header",
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Optional replay key. Canonical header form (x-idempotency-key is also supported)."
      },
...
      "DryRun": {
        "name": "dryRun",
        "in": "query",
        "description": "When true, analyze and return alerts without notification delivery, deduplication cache claims/writes, or signal-outcome persistence.",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
      "AlertListLimit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
...
      "SummaryLimit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000,
          "default": 500
        }
      },
      "ExportLimit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000,
          "default": 500
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Opaque pagination cursor returned by the previous page."
      },
      "Source": {
        "name": "source",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "webhook",
            "news-monitor",
            "market-scanner",
            "expanded-analysis"
          ]
        },
        "example": "webhook",
        "description": "Stored alert source filter. webhook covers /api/webhook/alert; news-monitor covers /api/news-monitor deliveries; market-scanner covers delivered /api/webhook/market-scanner-alert reports; expanded-analysis covers delivered /api/webhook/expanded-analysis-alert reports."
      },
      "Enriched": {
        "name": "enriched",
        "in": "query",
        "schema": {
          "type": "boolean"
        }
      },
      "IncludeText": {
        "name": "includeText",
        "in": "query",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "From": {
        "name": "from",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "To": {
        "name": "to",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "FromRequired": {
        "name": "from",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "ToRequired": {
        "name": "to",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "ExportFormat": {
        "name": "format",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "jsonl",
            "csv"
          ],
          "default": "jsonl"
        }
      },
...
      "XIdempotencyKeyHeader": {
        "name": "x-idempotency-key",
        "in": "header",
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Optional replay key. Alternative x-prefixed header variant."
      },
...
    },
    "requestBodies": {
...
      "Replay": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "channels": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "telegram",
                      "whatsapp",
                      "discord"
                    ]
                  }
                }
              }
            },
            "example": {
              "channels": [
                "telegram"
              ]
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Error": {
        "description": "Request or service error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Invalid request",
              "code": "INVALID_REQUEST"
            }
          }
        }
      },
      "IdempotencyConflict": {
        "description": "The idempotency key was reused with a different request fingerprint",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Idempotency key was reused with a different payload",
              "code": "IDEMPOTENCY_CONFLICT"
            }
          }
        }
      },
...
      "JsonResult": {
        "description": "Successful JSON response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JsonObject"
            }
          }
        }
      },
      "AlertSummaryResult": {
        "description": "Protected stored-alert analytics summary including bounded risk-metadata coverage and safe prompt provenance groups",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AlertSummaryResult"
            }
          }
        }
      },
...
      "DeliveryResult": {
        "description": "Per-channel delivery result. When ENABLE_ALERT_SIGNAL_REPEAT_SUPPRESSION is enabled and the same exchange/symbol/timeframe/side signal fires within its cooldown window, the response includes suppressedRepeat: true with empty results and deliveredChannels while the alert remains persisted with a suppression marker.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DeliveryResult"
            },
            "example": {
              "success": true,
              "requestId": "req-123",
              "results": [
                {
                  "channel": "telegram",
                  "success": true
                }
              ]
            }
          }
        }
      },
...
      "AlertList": {
        "description": "Stored alert page",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "alerts": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoredAlert"
                  }
                },
                "nextCursor": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "jobId": {
            "type": "string",
            "description": "Durably stored job identifier returned when queue acceptance is indeterminate."
          }
        }
      },
...
      "StoredAlert": {
        "type": "object",
        "description": "Stored alert document. The `source` field identifies the producing flow (webhook, news-monitor, market-scanner, expanded-analysis). When the report text exceeded the 20,000-character storage cap, `truncated: true` and `originalLength` are populated so consumers and replay can flag the loss; replay will redeliver only the truncated text.",
        "properties": {
          "id": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "text": {
            "type": "string"
          },
          "enriched": {
            "type": "boolean"
          },
          "enrichmentData": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "tokenUsage": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "channels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deliveryResults": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "source": {
            "type": "string",
            "enum": [
              "webhook",
              "news-monitor",
              "market-scanner",
              "expanded-analysis"
            ]
          },
          "useTradingViewData": {
            "type": "boolean"
          },
          "tradingViewEnrichmentApplied": {
            "type": "boolean"
          },
          "symbol": {
            "type": [
              "string",
              "null"
            ]
          },
          "exchange": {
            "type": [
              "string",
              "null"
            ]
          },
          "truncated": {
            "type": "boolean",
            "description": "True when the stored text was clipped to the 20,000-character cap; see originalLength for the pre-truncation size."
          },
          "originalLength": {
            "type": "integer",
            "minimum": 20001,
            "description": "Length of the original alert text in characters before the 20,000-character cap was applied."
          }
        }
      },
...
      "AlertSummaryResult": {
        "type": "object",
        "required": [
          "success",
          "summary"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "summary": {
            "$ref": "#/components/schemas/AlertSummary"
          }
        }
      },
...
    }
  }
}